Moved mediainfo to mnav app, updated tv app to use mediainfo
A new method has been added to the BaseVideoFile model - update_mediainfo()
This function will use mediainfo to update the information such as filesize, codec, etc

The TV app's management commands, tv_scan and tv_scanall has been updated to call this
when the mediainfo is out of date. The first time this is run it might take a while to complete.

mediainfo is considered out of data if the modify time on a filename is newer than the last time
the information was updated.